feat(step-4): connect LangChain4j to Claude via AnthropicChatModel (#5)
- Add LangChain4j BOM 1.10.0 + Anthropic/core/jdk-http deps to build.gradle
- Replace application.properties with application.yml (github/anthropic/mcp config)
- Create agent/BacklogAgent interface with @SystemMessage/@UserMessage
- Create config/LangChainConfig with AnthropicChatModel bean and AiServices builder
- Remove duplicate files with wrong packages (BacklogAgent.java, conf/LangChainConfig.java)
- Add src/test/resources/application-test.yml for CI without real API key
- Mock AnthropicChatModel and BacklogAgent in tests to avoid requiring real credentials
- Disable GitHubMcpToolsTest until STEP 5-6 classes are implemented
