--- src/main/java/com/example/agent/conf/LangChainConfig.java
@@ -1,59 +1,0 @@
-package net.filecode.agent.config;
-
-import dev.langchain4j.model.anthropic.AnthropicChatModel;
-//import dev.langchain4j.model.openai.OpenAiChatModel;
-import dev.langchain4j.service.AiServices;
-import java.time.Duration;
-import java.util.List;
-import net.filecode.agent.BacklogAgent;
-import net.filecode.agent.tools.AgentTool;
-import org.springframework.beans.factory.ObjectProvider;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class LangChainConfig {
-
-    @Bean
-    public AnthropicChatModel anthropicChatModel(
-        @Value("${anthropic.api-key}") String apiKey,
-        @Value("${anthropic.model}") String model,
-        @Value("${anthropic.max-tokens:800}") Integer maxTokens,
-        @Value("${anthropic.timeout-seconds:60}") Integer timeoutSeconds
-    ) {
-        return AnthropicChatModel.builder()
-            .apiKey(apiKey)
-            .modelName(model)
-            .maxTokens(maxTokens)
-            .timeout(Duration.ofSeconds(timeoutSeconds))
-            .build();
-    }
-
-    //  @Bean
-    //  public OpenAiChatModel openAiChatModel(
-    //          @Value("${openai.api-key}") String apiKey,
-    //          @Value("${openai.model}") String model,
-    //          @Value("${openai.timeout-seconds:60}") Integer timeoutSeconds
-    //  ) {
-    //    return OpenAiChatModel.builder()
-    //            .apiKey(apiKey)
-    //            .modelName(model) // gpt-4o-mini
-    //            .timeout(Duration.ofSeconds(timeoutSeconds))
-    //            .build();
-    //  }
-
-    @Bean
-    public BacklogAgent backlogAgent(
-        AnthropicChatModel model,
-        List<AgentTool> tools
-    ) {
-        System.out.println("=== Agent tools loaded: " + tools.size() + " ===");
-        tools.forEach(t -> System.out.println(" - " + t.getClass().getName()));
-
-        return AiServices.builder(BacklogAgent.class)
-            .chatModel(model)
-            .tools(tools.toArray())
-            .build();
-    }
-}
