+++ src/main/resources/application.yml
@@ -1,0 +1,16 @@
+spring:
+  application:
+    name: agent
+
+github:
+  owner: ${GITHUB_OWNER}
+  repo: ${GITHUB_REPO}
+
+anthropic:
+  api-key: ${ANTHROPIC_API_KEY}
+  model: claude-opus-4-20250514
+  timeout-seconds: 60
+
+mcp:
+  base-url: http://localhost:3333
+  path: /mcp
