+++ src/test/java/com/example/agent/tools/GitHubMcpToolsTest.java
@@ -1,0 +1,14 @@
+package com.example.agent.tools;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+
+// TODO STEP 8 — activate when McpHttpClient and GitHubMcpTools are implemented (STEP 5 & 6)
+@Disabled("Waiting for STEP 5 (McpHttpClient) and STEP 6 (GitHubMcpTools)")
+class GitHubMcpToolsTest {
+
+    @Test
+    void should_call_mcp_tool() {
+        // will be implemented in STEP 8
+    }
+}
