+++ src/test/resources/application-test.yml
@@ -1,0 +1,12 @@
+github:
+  owner: test-owner
+  repo: test-repo
+
+anthropic:
+  api-key: test-api-key
+  model: claude-opus-4-20250514
+  timeout-seconds: 10
+
+mcp:
+  base-url: http://localhost:3333
+  path: /mcp
